18 refactor to implement dao pattern

+++ Table-football-manager/src/main/resources/Data/init/players.sql
@@ -2,6 +2,6 @@ DROP TABLE IF EXISTS players CASCADE;
 CREATE TABLE players (
     id SERIAL PRIMARY KEY,
     name VARCHAR(255) UNIQUE NOT NULL,
-    team_id INT,
+    team_id BIGINT,
     CONSTRAINT fk_players FOREIGN KEY (team_id) REFERENCES teams (id) ON DELETE SET NULL
 )
